src/use-cases/users/authenticate.ts
import { UsersRepository } from '@/respositories/users-repository'
import { Consumer, User } from '@prisma/client'
import { compare } from 'bcryptjs'
import { InvalidCredentialsError } from '../errors/invalid-credentials-error'
import { MailValidationError } from '../errors/mail-validation-error'

interface AuthenticateUseCaseRequest {
  email: string
  password: string
}

interface AuthenticateUseCaseResponse {
  user: User
  consumer: Consumer | null
}

export class AuthenticateUseCase {
  constructor(private usersRepository: UsersRepository) {}

  async execute({
    email,
    password,
  }: AuthenticateUseCaseRequest): Promise<AuthenticateUseCaseResponse> {
    const user = await this.usersRepository.findByEmail(email)

    if (!user) {
      throw new InvalidCredentialsError()
    }

    const doesPasswordMatches = await compare(password, user.password_hash)

    if (!doesPasswordMatches) {
      throw new InvalidCredentialsError()
    }

    if (!user.mail_confirm_at) {
      throw new MailValidationError()
    }

    user.password_hash = 'omited'

    return { user, consumer: user.consumer }
  }
}
